requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
asyncio==3.4.3

# AI Provider APIs
openai==1.3.0
groq==0.4.0
google-generativeai==0.3.0

# CLI and utilities
click==8.1.7
rich==13.7.0
typer==0.9.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Optional dependencies
jupyter==1.0.0
pandas==2.1.3
numpy==1.25.2
beautifulsoup4==4.12.2
selenium==4.15.0
